% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidymicrobiome_update.R
\name{check_tidymicrobiome_version}
\alias{check_tidymicrobiome_version}
\title{check_tidymicrobiome_version}
\usage{
check_tidymicrobiome_version(
  packages = c("core", "all"),
  from = c("gitlab", "github", "shen")
)
}
\arguments{
\item{packages}{core or all packages in tidymicrobiome. "core" means all the core
packages in tidymicrobiome and "all" means all the packages in tidymicrobiome.}

\item{from}{"gitlab", "github", or "tidymicrobiome.org"}
}
\description{
Check if there are packages in tidymicrobiome can be updated.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
