% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-conflicts.R
\name{tidymicrobiome_conflicts}
\alias{tidymicrobiome_conflicts}
\title{Conflicts between the tidymicrobiome and other packages}
\usage{
tidymicrobiome_conflicts()
}
\description{
This function lists all the conflicts between packages in the tidymicrobiome
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
tidymicrobiome_conflicts()
}
