% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-tidymicrobiome_logo.R
\name{tidymicrobiome_logo}
\alias{tidymicrobiome_logo}
\title{Show the logo tidymicrobiome.}
\usage{
tidymicrobiome_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
A ASCII log of tidymicrobiome
}
\description{
The tidymicrobiome logo, using ASCII or Unicode characters
}
\examples{
tidymicrobiome_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
