% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidymicrobiome_update.R
\name{update_tidymicrobiome}
\alias{update_tidymicrobiome}
\title{update_tidymicrobiome}
\usage{
update_tidymicrobiome(
  packages = c("core", "all"),
  from = c("gitlab", "github", "tidymicrobiome.org"),
  fastgit = FALSE
)
}
\arguments{
\item{packages}{core or all packages in tidymicrobiome. "core" means all the core
packages in tidymicrobiome and "all" means all the packages in tidymicrobiome.}

\item{from}{github, gitlab or tidymicrobiome.org}

\item{fastgit}{if install packages using fastgit. see
https://hub.fastgit.org/}
}
\description{
Update packages in tidymicrobiome.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
